refactor: move files from bff folder to gql folder
